chore: release version 0.1.17 with enhancements and fixes
- Added ExceptionHelper for improved exception handling in various components.
- Enhanced SimVar type inference for automatic detection of SimConnectDataInitPosition.
- Improved struct-based SimVars.SetAsync<TStruct> for better performance and error handling.
- Updated logging mechanisms to use ExceptionHelper for critical exception filtering.
- Refined input event processing with defensive coding practices.
- Maintained integration tests for better coverage and reliability.
- Fixed issues with struct writers and SimVarManager logging.

+++ tests/SimConnect.NET.Tests.Net8/Tests/AIObjectTests.cs
@@ -124,10 +124,10 @@ namespace SimConnect.NET.Tests.Net8.Tests
                 {
                     var position = new SimConnectDataInitPosition
                     {
-                        Latitude = currentLat + (0.001 * (i + 1)),
-                        Longitude = currentLon + (0.001 * (i + 1)),
+                        Latitude = currentLat + (0.001 * (double)(i + 1)),
+                        Longitude = currentLon + (0.001 * (double)(i + 1)),
                         Altitude = 100,
-                        Heading = 90 + (i * 30),
+                        Heading = 90.0 + ((double)i * 30.0),
                         OnGround = 1,
                         Airspeed = 0,
                     };
